+++ lib/src/model_crud.dart
@@ -167,6 +167,24 @@ mixin ModelCrud implements ModelAccessor {
               as T);
 
   @override
+  Future<void> update<T extends ModelCrud>(
+          String id, Map<String, dynamic> updates) =>
+      ModelUtility.update<T>($models, $pathOf, updates, id);
+
+  @override
+  Future<void> setIfAbsent<T extends ModelCrud>(String id, T model) =>
+      exists(id).then((v) => v ? Future.value() : set<T>(id, model));
+
+  @override
+  Future<void> setIfAbsentUnique<T extends ModelCrud>(T model) =>
+      existsUnique<T>().then((v) => v ? Future.value() : setUnique<T>(model));
+
+  @override
+  Future<void> updateUnique<T extends ModelCrud>(
+          Map<String, dynamic> updates) =>
+      ModelUtility.update<T>($models, $pathOf, updates, null);
+
+  @override
   Future<void> deleteSelf<T extends ModelCrud>() {
     if (documentPath == null) {
       throw Exception("Cannot delete self without a document path");
@@ -181,6 +199,14 @@ mixin ModelCrud implements ModelAccessor {
   @override
   Future<bool> existsUnique<T extends ModelCrud>() =>
       getUnique<T>().then((value) => value != null).catchError((e) => false);
+
+  @override
+  Future<T?> getCached<T extends ModelCrud>(String id) =>
+      ModelUtility.pullCached<T>($models, $pathOf, id);
+
+  @override
+  Future<T?> getCachedUnique<T extends ModelCrud>() =>
+      ModelUtility.pullCached<T>($models, $pathOf, null);
 
   @override
   Future<void> setSelfAtomic<T extends ModelCrud>(T Function(T? data) txn) {